requirements.txt
streamlit
openai
requests
pandas
numpy
yfinance
python-dotenv
newspaper3k
plotly
